Add files via upload
Welcome to the repository for my personal portfolio website! This repository contains all the necessary files and code to showcase my skills, projects, and experiences in a visually appealing manner.

Features:

HTML: The portfolio.html file serves as the main entry point for the portfolio website. It provides the structure and layout of the different sections of the website.

CSS: The portfolio.css file contains the cascading style sheets that define the presentation and design of the website. It includes responsive design principles to ensure the website looks great on various devices and screen sizes.

Images: The .png files are the image assets used in the portfolio, including  profile picture, and other visual elements.

Projects: The portfolio showcases a variety of projects I have worked on. Each project is represented as a separate section with details about the project's description, technologies used, and a link to the project's live demo or repository.

About Me: The "About Me" section gives visitors an insight into my background, skills, and interests. It includes a brief biography and my contact information.

Contact: The "Contact" section provides a way for visitors to get in touch with me. It includes links to my social media profiles, email address, and possibly a contact form.

Usage:

Clone the repository to your local machine using git clone https://github.com/your-username/portfolio.git.
Open the portfolio.html file in your preferred web browser to view the portfolio website.
Customize the content in the HTML and CSS files to match your personal information and preferences.
Replace the images in the images directory with your own images or project screenshots.
Contributing:

Feel free to contribute to this repository by suggesting improvements, reporting issues, or adding new features. Fork the repository, create a new branch for your changes, and submit a pull request.